Add basic support for docstrings
This commit adds only one-line function description docstring support.

We can now do the following:

```
module mytool (foo, bar)

--' foo is a thing
foo :: A -> B

--' bar is also a thing
bar :: A -> B
```

The `--'` syntax is distinguished from a simple comment in the
lexer. The following string, up to the newline, is stored and passed as
a list of docstrings along with the following object.

When compiled, the docstring is propagated to the usage statement as so:

```
$ ./nexus
Usage: ./nexus [OPTION]... COMMAND [ARG]...

Nexus Options:
 -h, --help            Print this help message
 -o, --output-file     Print to this file instead of STDOUT
 -f, --output-format   Output format [json|mpk|voidstar]

Exported Commands:
  foo   foo is a thing
          param 1: A
          return: B
  bar   bar is also a thing
          param 1: A
          return: B
```

Here is what I would LIKE it to generate:

```
Exported Commands (-h for more info:
  foo   foo is a thing
  bar   bar is also a thing
```

Leave the details out of the top-level print session

```
$ mytool foo -h
Usage: mytool foo THINGA

foo is a thing

<details from additional docstrings>

Nexus Options:
 -h, --help            Print this help message
 -o, --output-file     Print to this file instead of STDOUT
 -f, --output-format   Output format [json|mpk|voidstar]

Positional Arguments:
 1) THINGA - Some thingy
    type: A

    <extra docs about thinga>

Return: some description of the return thing

<extra docs about return>

Return type: B
```

TODO:
 [ ] argument docstrings (partially supported)
 [ ] command usage option (./nexus foo -h)
 [ ] handling for single-command tools. For these, listing exported
     commands is redundant.
 [ ] propagate module name and use this as the name of the created executable
 [ ] properly install the executable in Morloc home
 [ ] add compiler option to uninstall a program