Add the multiprocessing test
Forgot to actually add it to the test suite earlier. Am I being stupid
here? Why don't I just treat every folder in the directory as a test and
automatically handle them? Originally, I wanted the golen tests to show a
description of what was being tested, but ultimately it was way more
useful to just show the directory name.

But maybe I can be even smarter. Maybe every test can have a bit of
metadata associated with it?

In any case, having to add the test and then update the Main.hs test
script is a pain.