Added namespace to Python pool
Now I have a more stable solution to module imports.

This is a breaking change. Previously I could role Python builtins into
source statements like so:

```
source Py from "foo.py" ("foo", "len")
```

Here `len` is a builtin function. Now the code generator would place a
namespace in front of `len`, just as it would with `foo`. Python will
have none of this and dies immediately, saying that the module doesn't
export `len`. This is correct. Indeed, it is sloppy and strange to even
write above the "foo.py" exported "len". It happened to be OK since the
generated code was a bit loose, but it was not a good idea.

Worse than this, I had been importing everything from all modules
(`from foo import *`), which is dreadful practice that can lead to
silent function masking.

I also need to fix this problem in R and C++ code. Currently the work,
since they depend on string paths to code, so the dash is no
problem. But the namespacing is still a concern there.