Make all bound names unique in Treeify (fix #37)
The problem described in #37 was simply caused by not substituting
lambda-bound terms when substituting a new variable definition in the
tree. These trees may be large, so I would like to avoid recursing up
and down them searching for free terms. So instead, I add a tag to teach
lambda-bound term ('@1', '@2', etc) that is syntactically illegal (so no
possibility of conflict with existing names) and autoincrementing. Every
time we see a lambda term, we update a map from existing names to unique
name (e.g., "x" -> "x@34").