Send args lambda function (fix #35)
Previously when a lambda was applied, but the lambda function contained
no CallS, for example `\ x f -> (x, f x)`, the implementation language
of the lambda function could not be inferred from the function
itself. It is dependent on an applied argument. Now I pass the state
needed to evaluate the language scores for bound terms (like `f` and `x`
above).

This fixes my test for #35, but more testing is needed before pushing to
master.