(2/437 fail) Refactor for undirected term search
I've been thrown for a bit of a loop by this latest issue. I want to
define implementations for functions when only the type signatures for
the term are available. For example:

```morloc
fst a b :: (a, b) -> a
snd a b :: (a, b) -> b

uncurry a b c :: (a -> b -> c) -> (a, b) -> c
uncurry f xy = f (fst xy) (snd xy)
```

This is from the new `root` module. Here `fst` and `snd` currently
must be sourced from foreign languages (this will change soon, but that
is a different problem). `uncurry`, though, can be defined in terms of
these two functions. So I want to define this pure Morloc form of
`uncurry` in the abstract, without relying on any particular
implementation of `fst` and `snd`. In this case, the `root` module
contains only type signatures and Morloc compositions -- nothing is
sourced from concrete languages. The language-specific modules, such as
`root-py` and `root-r`, will import `root` and supply any needed sourced
functions. But this means tha the `fst` and `snd` implementations in
`uncurry` need to be searched for in the modules that import
`root`. This is backwards of the normal case.

So when I merge all the information on every term in Treeify, I need to
search not just among the dependencies of a module, but also backwards
and sideways. I need to do an undirected neighbor search. In this
commit, I've added a new functions for neighbor searching to DAG and
also refactored Treeify to create a dedicated `LinkTerms` module where
all the merging info can be localized.

In the next commit, hopefully, I'll update the `linkTerms` function to
do this undirected search.