(1/437 fail) "Fix" neighbor traversal algorithm
The algorithm is still ugly as hell. It is a depth first neighbor
searching algorithm that tracks the terms that have been evaluated so
far in each module in order to avoid infinite cycles. This works, but it
loads dependencies "upside-down" often by traversing all the way down to
the most base dependencies and then walking up to the dependents and
back to origin. A breadth-first algorithm would be more natural.

Anyway, all tests pass except one. Fix this next (probably something
with aliasing) and then I can get back to writing and testing the root
library.