Check against type annotations (fix #47)
Now in the typechecker:

 1. every synth case looks up the annotation, and if found
    checks against it.

 2. every check case looks up the annotation, and if found
    subtypes the current type against the annotation type

Also, no only one type annotation is allowed. In the past, I allowed
many annotations because, at that time, I was writing explicit concrete
types. So a type annotation would require a concrete type in each
supported language (which was such a syntactically convoluted thing that
I never used it, hence why annotation support was missing).

I have provided one test for annotations (the motivating case from #47)
as `type-annotations-1`. This is the first in a series of tests that I
need to write to ensure annotations are properly handled.