(99/440 fail) Fix more bugs in the link code
The root library is nearly working now.

Given the new reverse-tree dependency tracing, it is now difficult to
specify which implementations are in scope. I may need some filtering
mechanism.

The failing tests are mostly due to the changes in import strategy. Now
it is an error to import a typeclass method explicitly. I need to go
through all the tests and convert them to the implementation and have
them use the new root module.

I may also need to write a test module for each language that exposes
monomorphic simple testing terms. If I use the new root, most of the
functions will be from typeclasses. Which, of course, is something that
needs to be tested as well.

I may need to rewrite my entire test suite from scratch someday. It has
become pretty klunky.