(19/440 fail) Fixed test cases for interop
There is a deep problem in the Generate:expressPolyExpr. The algorithm
is branching on equality between the number of lambda parameters and the
number of terms applied to the wrapped function. This is absolute
non-sense. There is no relation between the two. It worked for my simple
tests, but fails on anything beyond simple argument passing.

But before addressing this problem, I will fix all trivial problems in
my tests and then write a suite of tests that specifically check all
cases in expressPolyExpr.