Fixed the partial application problem ... I guess
Sooooo, like, I checked under the hood and saw where the problem
was. And I thought to myself, well, goshdarn it all, this looks
complicated. I could try to understand the twisted mind of the whoever
wrote this shit ... or I could just delete all the crap and see what
happens. And ... yeah ... deleting fixed the problem.

All tests pass, and those new ones too. What was that section for, again?

And, in order, my `expressPolyExpr` documentations lists the cases as:
 * case #4
 * case #8,
 * case ExpressPolyExpr Lam
 * case #0
 * case #2
 * case #1
 * case #1 (again),
 * case #5
 * case #2 (again)
 * case #6
 * express' VarS.

Maybe should straighten that out?

This section is has been blasted to hell and rewritten so many
times. Like something that crawled out of nature's garage, only alive
because of the compiler's selective pressure. Code being acted on
continuously by a mindless force. Ugh ... this mindless force needs more
coffee.