Allow general types to be declared and exported
Now we can write declarations such as:

```
type Str
type Foo a b
```

This may be exported and concrete forms may be associated with them. The
advantage is that now the root, fully general, library may be export
these types and they needn't be imported from dependent libraries that
provide implementations.