(1/441 fail) Temp fix for HKT (fix #51)
In types like

```morloc
morloc_map a b f :: (a -> b) -> f a -> f b
```

The qualifier `f` is a higher-kinded type. The compiler stores the
values of all qualifiers and uses these to explicitly declare template
parameters in C++. But C++ does not support higher-kinded types. So we
can write an expression like: `morloc_map<int,double>`, but not
`morloc_map<int,double,vector>`.

As a temporary fix, I simply ignore the higher-kinded types. This will
work for C++ if the higher kinded cases are resolved through overloading
or by seperate functions pointed to by the typeclass instances. Then,
for the map example, only the `a` and `b` parameters would be declared.

All of this is rather hacky, but I'm not sure of a better solution yet.