Fix XDG handling for basic imports
Now:

 * morloc home: ./local/share/morloc
 * morloc lib: ./local/share/morloc/src/morloc/plane/default/

Why so deep?
 * ./local/share/morloc  -- this is the XDG spec approved local
 * ./local/share/morloc/src  -- may contain more than just morloc source
 * ./local/share/morloc/src/morloc/  -- may contain extra-planar contend
 * ./local/share/morloc/src/morloc/plane  -- may be many planes
 * ./local/share/morloc/src/morloc/plane/default  -- for now, this is one

There is actually one more layer in the users
directory: ./local/share/morloc/versions. This contains all the
different installed morloc versions as well as the "local" version for
local morloc builds. When morloc is run in a container, the particular
version is mounted at ./local/share/morloc.