Allow underscores in variable names
Previously I forced variables to use the Haskell convention of
alphanumeric plus quotes. But maybe it is better to let the people
decide. Some of the main functions that users will add are large
specialized functions that should, perhaps, have descriptive names. And
longer names are more readable in snakecase. However, camel case is more
tightly coupled, it seperates the function names more cleanly as
individual tokens from the surrounding symbolic clutter. Which is
better?