Remove --fast flag from stack build
The morloc compiler has hits some VERY bad performance in certain cases,
such as lists of expressions. I'm not sure of the root problem yet. But
to this point I've not thought much about performance of the compiler
itself.

But compiling without the --fast flag (which turns on most GHC
optimizations) improves performance ~7X (from 25s to 3.5s for the hofs-1
test). I need to start profiling my code and hunting down the
problematic sections.