Add string interpolation
This works with basic test coverage over R, Python and C++. It still
needs more thorough testing.

In the pure Morloc case, string interpolation is problematic. The reason
is that the nexus does not have a real code interpreter. Rather it
simpler inserts arguments into generated strings. The arguments are not
evaluated. For string interpolation, this screws up quotation.

What I need is a real code interpreter that translates incoming
arguments to voidstar and then actually executes the morloc code.