Add get pattern type checking
This required some fairly deep changes to the typechecker. The patterns
generate existential types that store the information about the expected
input type that can be inferred from the pattern alone. For tuple
access, such as .2, the given value must be a tuple and it must have at
least 3 elements, but it may have more. This generates an existential
type with 3 elements. When this is subtyped against another type, no
error will be raised if the other types has more than three elements,
rather the existential will expand. Same with records. But not all
existentials are "open" like this. So now existentials have an extra bit
of info for the parameters and keyword lists that store if they are
"open" or "closed" sets.

TODO:
 [ ] much more getter testing
 [ ] add R and C++ support
 [ ] add pure-morloc support (will need to write the nexus interpreter)