Setter patterns working for Python
I wrote a dedicated function in the pool that takes a setter pattern and
applies it to the data. It is not particularly efficient. I ought to
instead handle pattern evaluation in the Morloc clib.

Also, I replaced all the old ugly MT.Text with just Text using a direct
import of the Text type.