(6/444 fail) Add flawed pure interpretation
Previously, very simple pure Morloc programs could be executed that
basically just read data and substituted in arguments. This allowed
trivial things like `foo x = [x]` to be written and tested. The inputs
were not validated or parsed in any way. This was hacky and basically
just used to make my tests pass and for little demos. But with the
introduction of patterns, what is possible in pure Morloc is greatly
expanded. So we need a true pure interpreter. This commit is a first
step towards that.

The interpreter will read input arguments in the same way as a normal
command, with validation and support for all input formats (JSON,
MessagePack, and voidstar). The interpreter will further be able to
apply functions (variable substitution) and execute getter and setter
patterns. Eventually I will also add a small set of builtin functions
with plugin support.