All tests pass for pure morloc interpretter
There is a bit of a change here, now pure Morloc code is strict and
requires full type definitions. Previously code like this was possible:

```
foo x = (x, 32, True)
```

The value of `x`, whatever JSON data the user provided, was substituted
in. This was a janky-as-hell form of polymorphism that isn't OK in any
universe. Eventually, it would be cool to be able to write polymorphic
programs that infer input types and concretize accordingly.

So here's the TODO list:
 * Immediate
   [ ] add getter/setter pure morloc support
   [ ] add more comprehensive pure morloc tests
 * Future
   [ ] add input polymorphism