Fix parsing of docstrings
Docstrings always describe the element that follows them. There is no
need for them to have precise whitespace alignment. So now they are not
required to have any particular indentation.

It took me a devilishly long time to fix the parser for docstrings. No
thanks to the expensive misdirections from Claude.