Add usage statements for argument unrolling
This is proving to be a lot more complex than I had anticipated. There
are many subjective choices to make in rendering the usage
statements. For each argument, I need to render four things:
  1. the argument name (`-s, --long` for optional args, metavar or
     something for positional)
  2. the one-line docstring
  3. the type
  4. the default

But what about the detailed additional lines after the one line
docstring? And what if the type or default is too long for one line?
What about flags?

For now, I have almost all the featues done. I can tweak it
later. Beyond style, what is missing is just flags and defaults.

I'll circle back to this, but next I want to get the full argument
handling working.