Fix flag handling for generated CLI
The logic all works, a remaining problem is that I do not print the
default values to the help.

I could write the default after the argument, like so:

```
    -t INT, --threads=INT    default: 1
        number of threads to use
    --clean    default: false
        clean out caches before running?
```

Or on a dedicated line:

```
    -t INT, --threads=INT
        number of threads to use
        default: 1
    --clean
        clean out caches before running?
        default: false
```

I also need to add type:

```
    -t INT, --threads=INT
        number of threads to use
        type: Int
        default: 1
    --clean
        clean out caches before running?
        type: Bool
        default: false
```

Too cluttered? Maybe type doesn't need to be added for flags or for
primitives (ints, floats, strings) where the type can be assumed or
encoded in the metavar?

Also descriptions can be longer, so maybe the fields of metadata should
go first?

```
    -t INT, --threads=INT
        default: 1
        number of threads to use

    --clean
        default: false
        clean out caches before running?
```

Or should flags ALWAYS be false by default and flag presence toggles
them on? Or should should flag arguments have both? As in a `--clean`
and a `--no-clean` argument? This would be helpful when we are loading a
custom default record. If these options were added, then the default
would be made illegal for booleans. This could then make flags much
easier to work with since the presence always means true, rather than
just the inverse of whateve the default happens to be.

```
    --clean
        clean out caches before running?

    --no-clean
        inverse of --clean
```

Or with short args:

```
    -c, --clean
        clean out caches before running?

        --no-clean
        inverse of --clean
```

And these could be specified in morloc code like so:

```
--' arg: -c/--clean
--' arg-off: -C/--no-clean
```

Should the reverse flags also have their own docstrings?

Or more commonly:

```
--' arg: -c/--clean
```