Fix dual option flags
Some flags are associated with two options. One sets the value to the
opposite of the default and the other sets the value to the default. Now
I support both of these under the "true:" and "false:" docstrings.

This commit represents a huge refactor of the docstring handling
code. All logic for determining docstrings types is now localized in the
Docstrings module. The types are now much tighter and all the prior
error cases are now caught by the type system.