Add docstring return tag
This is isn't strictly necessary, since we could just place a docstring
on the last entry in the signature, but it is helpful for simple
signatures where none of the fields have explicit annotations. For
example:

```
--' return: size in bytes, or 0 if file cannot be accessed
getFileSize :: Filename -> Int
```

Versus

```
getFileSize :: Filename ->
  --' size in bytes, or 0 if file cannot be accessed
  Int
```

Is this added tag type worth it? As currently implemented, the `return`
flag over-rides any final argument docstring. For example:

```
--' return: a
foo ::
  Int ->
  --' b
  Bool

The return description will be "a"