Infer positional if no default given
There are cases where one might want to specify the options for a type
but not the default. In these cases, the user would be assuming that a
the default would be provided later. For example:

```
--' arg: --foo
type Foo = Int

bar ::
  --' default: 5
  Foo ->
  Str
```

A more common case would be for a record:

```
--' arg: --person-config
record Person where
  ...

-- Person is positional
f1 :: Person -> Int

-- Person is unrolled
f2 ::
  --' unroll: true
  Person ->
  Int

-- Person is an optional record
f2 ::
  --' unroll: false
  --' default: '{"name":"Weena", "age":23}'
  Person ->
  Int
```